+++ src/acuvity/response/parser.py
@@ -1,10 +1,9 @@
 from enum import Enum, auto
 from typing import Optional, Tuple, Union
 
-from ..guard.constants import GuardName
-from ..models.extraction import Extraction
-from ..guard.errors import ValidationError
-
+from acuvity.guard.constants import GuardName
+from acuvity.models.extraction import Extraction
+from acuvity.guard.errors import ValidationError
 
 class GuardType(Enum):
     """Types of guards and their corresponding sections in Extraction."""
